requirements.txt
datasets==2.19.0
pandas==2.2.2
requests==2.31.0
beautifulsoup4==4.12.3
pyyaml==6.0.1
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.0
joblib==1.4.2
jupyter==1.0.0
nbformat==5.10.4
nbconvert==7.16.0
pyarrow==16.0.0
nltk==3.8.1
wordcloud==1.9.3
anthropic==0.25.0
python-dotenv==1.0.1
streamlit==1.34.0
